+++ src/components/MovieList_Item/ButtonsHover/index.js
@@ -1,0 +1,3 @@
+import ButtonsHover from './ButtonHover';
+
+export default ButtonsHover;
