+++ src/components/MovieList_Item/MovieListItemMainInfo/MovieListItemMainInfo.jsx
