+++ src/components/MovieList_Item/MovieListItemMainInfo/index.js
@@ -1,0 +1,3 @@
+import MainInfo from './MovieListItemMainInfo';
+
+export default MainInfo;
