+++ src/components/MovieList_Item/MovieListItemMainInfo/mainInfo.scss
