+++ src/components/MovieList_Item/Overwiew/index.js
@@ -1,0 +1,3 @@
+import Overview from './Overview';
+
+export default Overview;
