+++ src/components/MovieList_Item/Trailer/index.js
@@ -1,0 +1,3 @@
+import Trailer from './Trailer.jsx';
+
+export default  Trailer;
