+++ src/components/MovieList_Item/index.js
@@ -1,0 +1,3 @@
+import MovieListItem from './MovieListItem';
+
+export default MovieListItem;
