+++ src/components/MovieList_Item/movieListItem.scss
