--- src/components/MovieList__Item/style.scss
@@ -1,184 +1,0 @@
-.movieList__item{
-  width: 20%;
-  min-width: 400px;
-  height: 600px;
-  margin-bottom: 25px;
-  background-color: white;
-}
-.movieList__item_bg{
-  position: relative;
-  box-sizing: border-box;
-  width: 100%;
-  height: 75%;
-  margin-bottom: 10px;
-  background-size: cover;
-}
-.movieList__item__mainInfo{
-  position: relative;
-}
-.movieList__item_genresList li{
-  display: inline-block;
-  margin-right: 5px;
-  color:  #47c3f0;
-  font-size: 18px;
-  letter-spacing: 1px;
-  line-height: 33px;
-  color: #47c3f0;
-  font-family: "Helvetica Neue";
-  font-weight: 500;
-}
-.movieList__item_mark{
-  position: absolute;
-  display: inline-block;
-  box-sizing: border-box;
-  width: 60px;
-  height: 36px;
-  text-align: center;
-  line-height: 36px;
-  border: 2px solid #47c3f0;
-  border-radius: 5px;
-  font-weight: bold;
-  color:  #47c3f0;
-}
-.movieList__item_title{
-  display: inline-block;
-  width: 70%;
-  margin-right: 10%;
-  font-size: 28px;
-  letter-spacing: 1px;
-  line-height: 44px;
-  font-family: "Helvetica Neue";
-  font-weight: bold;
-  white-space: nowrap;
-  overflow: hidden; 
-  text-overflow: ellipsis;
-}
-
-// styles for hover 
-
-.movieList__item_buttons_container{
-  display: none;
-}
-.movieList__item_buttons_container_show{
-  position: absolute;
-  top: 30%;
-  left: 25%;
-  display: block;
-
-}
-.movieList__item__button_watch_container{
-  width: 200px;
-  margin: 0 auto;
-}
-.movieList__item__button_watch{
-  position: relative;
-  width: 75px;
-  height: 75px;
-  margin-left: 63px;
-  background-color: white;
-}
-.movieList__item__button_watch::after{
-  position: absolute;
-  top: -25px;;
-  left: -25px;
-  font-family: "Font Awesome 5 Free";
-  font-size: 120px;
-  font-weight: 900;
-  content: "\f144";
-  color:  rgb(9, 188, 201);
-}
-.movieList__item__button_watch_container p{
-  margin-top: 30px;
-  font-size: 28px;
-  letter-spacing: 1px;
-  line-height: 44px;
-  color: #ffffff;
-  font-family: "Helvetica Neue";
-  text-align: center;
-}
-.movieList__item__button_view{
-  display: block;
-  height: 60px;
-  width: 199px;
-  margin: 50px auto;
-  color: white;
-  border: 1px solid white;
-  border-radius: 5px;
-  background-color: rgba(0, 0, 0, 0.7);
-}
-
-// styles for overview
-
-.movieList__item__overview{
-  display: none;
-}
-.movieList__item__overviewShow{
-  display: block;
-  box-sizing: border-box;
-  padding: 75px 0 0 40px;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.7);
-}
-.movieList__item__closeOverview{
-  width: 30px;
-  height: 30px;
-  text-align: center;
-  line-height: 30px;
-  background-color: rgba(0, 0, 0, 0);
-  border-style: none;
-  outline: none;
-  color: white;
-}
-.movieList__item__closeOverview:active{
-  color: rgb(105, 103, 103);
-}
-.movieList__item__closeOverview .fa-times{
-  font-size: 25px;
-}
-.movieList__item_buttons_container_hideWithButton{
-  display: none;
-}
-.movieList__item__overview{
-  color: white;
-}
-.movieList__item__overviewDetails{
-  max-height: 270px;
-  font-size: 20px;
-  line-height: 30px;
-  color: #ffffff;
-  font-family: "Helvetica Neue";
-  overflow: hidden; 
-  text-overflow: ellipsis;
-}
-.movieList__item_bg_overview{
-  height: 100%;
-}
-.movieList__item__mainInfoHide{
-  display: none;
-}
-.overview__watch{
-  position: absolute;
-  bottom: 75px;
-}
-
-// trailer
-
-.movieList__item__trailerContainer{
-  position: relative;
-  display: none;
-  width: 1200px;
-  height: 900px;
-  margin: 0 auto;
-  border: 1px solid grey;
-  background-color:rgb(216, 216, 216);
-  z-index: 99;
-}
-.movieList__item__trailerContainerShow{
-  display: block;
-}
-.movieList__item__trailer{
- position: absolute;
- top: 70px;
- left: 60px;
- z-index: 100;
-}
